DDC: Added ECachePolicy::KeepAlive and EBuildPolicy::CacheKeepAlive
The new flags are meant to prevent a cache store from evicting the associated cache records for the duration of the session.

#rb Zousar.Shaker
#rnx
#preflight 614954b611925a00017c525f

#ROBOMERGE-AUTHOR: devin.doucette
#ROBOMERGE-SOURCE: CL 17579887 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v871-17566257)

[CL 17579899 by devin doucette in ue5-release-engine-test branch]

+++ Engine/Source/Developer/DerivedDataCache/Private/DerivedDataBuildJobContext.h
@@ -18,8 +18,8 @@ namespace UE::DerivedData { class FBuildOutputBuilder; }
 namespace UE::DerivedData { class FPayload; }
 namespace UE::DerivedData { class IBuildJob; }
 namespace UE::DerivedData { class IRequestOwner; }
-namespace UE::DerivedData { enum class EBuildPolicy : uint8; }
-namespace UE::DerivedData { enum class ECachePolicy : uint8; }
+namespace UE::DerivedData { enum class EBuildPolicy : uint32; }
+namespace UE::DerivedData { enum class ECachePolicy : uint32; }
 namespace UE::DerivedData { enum class EPriority : uint8; }
 
 namespace UE::DerivedData::Private
